Switch git submodules to subtree merge
Git submoules works fine locally to link to the BurrTools code, but it
breaks on GitHub. Subtree mergingi is an alternate strategy which works
with GitHub's linking. The downside is that we have all of the BurrTools
code actually checked into this repository.

See the git book for details on subtree merge

    https://git-scm.com/book/en/v2/Git-Tools-Advanced-Merging#_subtree_merge